--- woo-ah-hana-web/src/app/business/memory/memory.d.ts
@@ -1,16 +1,0 @@
-export interface PaymentLogType {
-    tran_date: string;
-    tran_time: string;
-    inout_type: string;
-    tran_type: string;
-    print_content: string;
-    tran_amt: string;
-    after_balance_amt: string;
-    branch_name: string;
-}
-
-export interface GetPlanReceiptDto {
-    records: LogDataType[];
-    totalAmt: number;       
-    perAmt: number;  
-}
